### 2023-10-15 #5953 @kdm9
FYI: cloc doesn't count the actual code of Acanthophis, which is in snakemake `*.rules` files.
