### 2023-10-15 a49cc4c
only 2 cpus on test runners
