+++ tests/test.sh
@@ -6,4 +6,4 @@ acanthophis-init --yes
 mamba env update -f environment.yml
 conda activate acanthophis
 rm -fr output tmp
-snakemake -j 8 --use-conda --ri "${@}"
+snakemake -j 2 --use-conda --ri "${@}"
