### 2023-10-18 #5953 @marcosvital
Hi, @kdm9, I will be the editor of your submission.

Please take a look at the manuscript proof you generated above, and check if it seems ok and ready to review. If you make any changes or corrections on it you can create a new version anytime. I saw a small type: Stage 6 is written as Stage 5.

At the same time, we can start assigning reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this - you can paste the github profile of anyone you think that could do it (but please don't tag anyone yet).
You can also check and search using [this list of people](https://reviewers.joss.theoj.org/lookup) that have agreed to review for JOSS, and see if any of them may be suitable to review this submission.
