+++ paper/paper.bib
@@ -388,6 +388,27 @@
   file = {/home/kevin/work/bibliography/pdfs/ondov16_mash__genome_biology__mash.pdf}
 }
 
+@article{pedersen18_mosdepthquick,
+  title = {Mosdepth: Quick Coverage Calculation for Genomes and Exomes},
+  shorttitle = {Mosdepth},
+  author = {Pedersen, Brent S. and Quinlan, Aaron R.},
+  date = {2018-03-01},
+  journaltitle = {Bioinformatics (Oxford, England)},
+  shortjournal = {Bioinformatics},
+  volume = {34},
+  number = {5},
+  eprint = {29096012},
+  eprinttype = {pmid},
+  pages = {867--868},
+  issn = {1367-4811},
+  doi = {10.1093/bioinformatics/btx699},
+  abstract = {SUMMARY: Mosdepth is a new command-line tool for rapidly calculating genome-wide sequencing coverage. It measures depth from BAM or CRAM files at either each nucleotide position in a genome or for sets of genomic regions. Genomic regions may be specified as either a BED file to evaluate coverage across capture regions, or as a fixed-size window as required for copy-number calling. Mosdepth uses a simple algorithm that is computationally efficient and enables it to quickly produce coverage summaries. We demonstrate that mosdepth is faster than existing tools and provides flexibility in the types of coverage profiles produced. AVAILABILITY AND IMPLEMENTATION: mosdepth is available from https://github.com/brentp/mosdepth under the MIT license. CONTACT: bpederse@gmail.com. SUPPLEMENTARY INFORMATION: Supplementary data are available at Bioinformatics online.},
+  langid = {english},
+  pmcid = {PMC6030888},
+  keywords = {Algorithms,Exome Sequencing,{Genome, Human},Genomics,High-Throughput Nucleotide Sequencing,Humans,{Sequence Analysis, DNA},Software},
+  file = {/home/kevin/work/bibliography/pdfs/pedersen18_mosdepthquick__bioinformatics_(oxford,_england)__mosdepth.pdf;/home/kevin/work/bibliography/zotdir/storage/6NAYGY2D/Pedersen and Quinlan - 2018 - Mosdepth quick coverage calculation for genomes a.pdf}
+}
+
 @article{regalado20_combining,
   title = {Combining Whole-Genome Shotgun Sequencing and {{rRNA}} Gene Amplicon Analyses to Improve Detection of Microbe–Microbe Interaction Networks in Plant Leaves},
   author = {Regalado, Julian and Lundberg, Derek S. and Deusch, Oliver and Kersten, Sonja and Karasov, Talia and Poersch, Karin and Shirsekar, Gautam and Weigel, Detlef},
