### 2023-10-19 #5953 @kdm9
Hi @marcosvital , thanks for editing!

I have fixed those and a few extra minor grammar/citation issues, and hopefully now after a

the PDF should be in a final draft state and ready to review.

Regarding reviewers, from that list I would suggest the following, if they have time:

- https://github.com/aboffin
- https://github.com/gbouras13
- https://github.com/bricoletc
- https://github.com/standage
- https://github.com/camillescott
- https://github.com/conradstack

(though I note a bit of a lack of plant genomics/informatics folks in JOSS -- I need to do some spruiking of JOSS in these circles it seems!!)
