### 2023-10-19 81c6984
more paper typos & fixes
