### 2023-10-31 #5953 @marcosvital
> sorry I just tried this as someone did it on my paper being reviewed at the moment - apologies.

No problem at all, this command will work once we start the review (which will happen in a different GitHub issue).
