### 2023-11-07 #5953 @aboffin
@marcosvital<https://github.com/marcosvital> Thank you for the invite, unfortunately I am unable to review this submission at this moment.

From: Marcos Vital (LEQ-UFAL) ***@***.***>
Date: Thursday, October 26, 2023 at 9:02 AM
Cc: Senthil Murugapiran ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: Acanthophis: a comprehensive plant hologenomics pipeline (Issue #5953)

CAUTION: This email originated from outside of the organization. Do not click links or open attachments unless you recognize the sender and know the content is safe.

Dear @aboffin<https://github.com/aboffin>, @gbouras13<https://github.com/gbouras13> and @bricoletc<https://github.com/bricoletc>, would you be willing to review this submission for JOSS?

We carry out an open checklist-driven reviews here in GitHub issues, and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is still a pre-review issue, and once there are enough reviewers we will move on to a new issue where the review will take place. If you are not able to review, please let us know if you have recommendations for other reviewers that could help with this submission.
