### 2023-11-12 #5953 @abhishektiwari
@marcosvital I have used Snakemake pipelines, so open to review this paper. Feel free to assign me as a reviewer.
