### 2023-11-15 #6062 @editorialbot
[REVIEW]: Acanthophis: a comprehensive plant hologenomics pipeline
**Submitting author:** @kdm9 (<a href="http://orcid.org/0000-0002-2466-1917">Kevin Murray</a>)
**Repository:** https://github.com/kdm9/Acanthophis
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
