### 2023-11-30 kdm9/Acanthophis#7 @bricoletc
Unable to run tests: missing 'populate_metadata' in module 'acanthophis'
(Part of review thread https://github.com/openjournals/joss-reviews/issues/6062)

Hello! I tried running the tests for this repo by following the instructions [here](https://github.com/kdm9/Acanthophis/blob/main/tests/README.md) and found the following minor issues:

- The conda environment file is currently not `conda.yml` as per the README, but `environment-setup.yml`
- In that `environment-setup.yml`, installing msprime from pip failed - see attached log file
[failed_pip_install_msprime.txt](https://github.com/kdm9/Acanthophis/files/13512718/failed_pip_install_msprime.txt). I found that simply installing msprime from conda-forge worked instead, as per [instructions here](https://tskit.dev/msprime/docs/latest/installation.html).
  --> Update the .yml accordingly?
- The last part of the README in the `tests` directory should be updated to reflect the current status of Acanthophis - i.e.,
  using the instructions from [here](https://github.com/kdm9/Acanthophis#documentation) in the main README, or just link to that.
  Same for the comment '# once I put this on PyPI:' - I see acanthophis is on PyPI

And here is a **more major issue**:

-  Running `snakemake --snakefile Snakefile.generate-rawdata -j 8 --use-conda` completed fine and produced the `rawdata`. However, for actually running acanthophis on the generated
  test dataset, `snakemake --snakefile Snakefile -j 8 --use-conda`, `Snakefile` and `config.yml` are missing.
  I saw that inside the script `test.sh` there are instructions to generate them, using `acanthophis-init`. Please either
  tell the user to run `setup.sh` and `test.sh`, or `acanthophis-init`, in the README code block.
- Having done as above, command `snakemake --snakefile Snakefile -j 8 --use-conda` gives the following error:
  ```sh
  AttributeError in file /home/adminbrice/Desktop/research/writing/reviewing_exernal/JOSS/Acanthophis/test_tool/Acanthophis/tests/Snakefile, line 4:
    module 'acanthophis' has no attribute 'populate_metadata'
  File "/home/adminbrice/Desktop/research/writing/reviewing_exernal/JOSS/Acanthophis/test_tool/Acanthophis/tests/Snakefile", line 4, in <module>
  ```
