### 2023-11-30 kdm9/Acanthophis#8 @bricoletc
Missing contribution/community guidelines
(Part of review thread https://github.com/openjournals/joss-reviews/issues/6062)

The three following guidelines are missing, on how to: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

(Presumably points 2 and 3 can just be raising issues on the repo)
Please provide these either in a section in the main README or in a separate `CONTRIBUTING.md` file linked to in the README
