### 2023-11-30 #6062 @bricoletc
@kdm9, I've provided my main comments in three issues on your repo; I'll wait until they're addressed before checking in the corresponding items in my checklist

@marcosvital, I have two questions about JOSS: 1) why are there no 'Author contributions' sections in JOSS papers? This would help address checklist point 'contribution and authorship' 2) Can there be a 'Discussion' section in JOSS papers? I've provided two comments to that effect in https://github.com/kdm9/Acanthophis/issues/9 ; but is fine if not - just curious why they don't exist in JOSS papers
