### 2023-12-20 #6062 @marcosvital
> @marcosvital, I have two questions about JOSS: 1) why are there no 'Author contributions' sections in JOSS papers? This would help address checklist point 'contribution and authorship' 2) Can there be a 'Discussion' section in JOSS papers? I've provided two comments to that effect in [kdm9/Acanthophis#9](https://github.com/kdm9/Acanthophis/issues/9) ; but is fine if not - just curious why they don't exist in JOSS papers

Sorry about the delayed reply.

About a space for Author contributions, I'll take that to our editorial team so we can discuss that. Feels like a good suggestions to me.

About a Discussion section: it can be included - beside the mandatory sections from all papers, the authors can include specific sections that are needed for their work. We usually won't find a Discussion section in a JOSS paper because the journal is devoted to publishing the research software, but not the research findings.
