### 2024-01-09 kdm9/Acanthophis#10 @gbouras13
[JOSS Review] Installation Instructions in tests wrong .yml file
Re https://github.com/openjournals/joss-reviews/issues/6062

I am running the example deployment of Acanthophis based on the Readme in the `tests` directory.

The instruction is

`conda env create -f conda.yml -n acanthophis-demo`

but presumably the correct command should be

`conda env create -f environment-setup.yml -n acanthophis-demo`
