### 2024-01-09 kdm9/Acanthophis#11 @gbouras13
[JOSS Review] Improve `tests` docs
Re https://github.com/openjournals/joss-reviews/issues/6062

After running

`snakemake --snakefile Snakefile.generate-rawdata -j 8 --use-conda`

which completed perfectly, the instructions say:

"To initialise your own workflow, copy ./Snakefile and ./config.yml to a new directory"

However, there isn't a Snakefile or config.yml file in the tests directory. I assume that you need to start a new project as per the main documentation - it would be good to make this clearer.

George
