### 2024-01-09 kdm9/Acanthophis#12 @gbouras13
[JOSS Review] tab/space bug
Re https://github.com/openjournals/joss-reviews/issues/6062

After #11 copying the Snakefile, I ran:

```
snakemake --snakefile Snakefile -j 8 --use-conda
```

which yielded an error (at the bottom of this issue).

It can be fixed by replacing tab with spaces in the line

`        "   --prob-contamination 1e-3"`

in the freebayes rule in the varcall.rules file.

```
Traceback (most recent call last):
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/workflow.py", line 1369, in include
    exec(compile(code, snakefile.get_path_or_uri(), "exec"), self.globals)
  File "/Users/a1667917/Documents/Reviews/acanthophis/Acanthophis/tests/Snakefile", line 18, in <module>
    rules.varcall.input,
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/workflow.py", line 1347, in include
    code, linemap, rulecount = parse(
                               ^^^^^^
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 1307, in parse
    for t, orig_token in automaton.consume():
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 110, in consume
    for t, orig in self.state(token):
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 1212, in python
    for t in self.subautomaton(token.string, token=token).consume():
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 110, in consume
    for t, orig in self.state(token):
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 199, in block
    yield from self.block_content(token)
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 784, in block_content
    for t in self.block(e.token):
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 199, in block
    yield from self.block_content(token)
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 784, in block_content
    for t in self.block(e.token):
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 199, in block
    yield from self.block_content(token)
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 784, in block_content
    for t in self.block(e.token):
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 199, in block
    yield from self.block_content(token)
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 773, in block_content
    for t in self.subautomaton(
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 107, in consume
    for token in self.snakefile:
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/parser.py", line 1238, in __next__
    return next(self.tokens)
           ^^^^^^^^^^^^^^^^^
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/tokenize.py", line 541, in _generate_tokens_from_c_tokenizer
    raise e from None
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/tokenize.py", line 537, in _generate_tokens_from_c_tokenizer
    for info in it:
  File "<string>", line 101
    	"   --skip-coverage 100000"
    	                           ^
TabError: inconsistent use of tabs and spaces in indentation

Traceback (most recent call last):
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/cli.py", line 1887, in args_to_api
    dag_api = workflow_api.dag(
              ^^^^^^^^^^^^^^^^^
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/api.py", line 326, in dag
    return DAGApi(
           ^^^^^^^
  File "<string>", line 6, in __init__
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/api.py", line 421, in __post_init__
    self.workflow_api._workflow.dag_settings = self.dag_settings
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/api.py", line 372, in _workflow
    workflow.include(
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/workflow.py", line 1371, in include
    e = update_lineno(e, self.linemaps)
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/a1667917/mambaforge/envs/someproject/lib/python3.12/site-packages/snakemake/exceptions.py", line 165, in update_lineno
    ex.lineno = linemaps[ex.filename][ex.lineno]
                ~~~~~~~~^^^^^^^^^^^^^
KeyError: '<string>'

```
