### 2024-01-09 kdm9/Acanthophis#14 @gbouras13
[JOSS Review] Source install v1.0.0rc4.dev6 error on OSX
Re https://github.com/openjournals/joss-reviews/issues/6062

After #13 , I instead proceeded with a source install instead of pip

i.e.

```

mamba create -n someproject python snakemake pip natsort
mamba activate someproject
git clone https://github.com/kdm9/Acanthophis.git

cd Acanthophis
pip install -e .
```

which worked

```
Installing collected packages: acanthophis
Successfully installed acanthophis-1.0.0rc4.dev6
```

Then I tried to run `acanthophis-init`, which did not:

```
cd ..
acanthophis-init some_project
```

```
Traceback (most recent call last):
  File "/Users/a1667917/mambaforge/envs/someproject/bin/acanthophis-init", line 8, in <module>
    sys.exit(init())
             ^^^^^^
  File "/Users/a1667917/Documents/Reviews/acanthophis/Acanthophis/acanthophis/cmd.py", line 64, in init
    template_dir = acanthophis.get_resource("template/")
                   ^^^^^^^^^^^^^^^^^^^^^^^^
AttributeError: module 'acanthophis' has no attribute 'get_resource'
```

George
