### 2024-01-09 #6062 @gbouras13
Hi @kdm9 ,

Thanks for the bugging, much needed for me :)

I've left a few issues in the repo. I tried to install Acanthophis to test it out, but I am struggling to, either from source or from pip, so I cannot proceed on my review further.

I appreciate that it's been a few months since you submitted, so the documentation is probably out of date as it has taken me so long to review it.

But as an overall comment, I think Acanthophis would greatly benefit if you beefed up and made the README.md instructions in the `tests` repository part of the main documentation (or linked them on the main README.md). It would give users an example of how to set up Acanthophis and run a test dataset.

George
