### 2024-01-09 c520d34
Fix conda env path (closes #10)
