### 2024-01-09 d05a9cd
improve test docs to link to real docs, closes #11
