+++ tests/README.md
@@ -17,29 +17,25 @@ conda env create -f environment-setup.yml -n acanthophis-demo
 conda activate acanthophis-demo
 # Install Acanthophis
 pip install -e ..
-# once I put this on PyPI:
-# pip install acanthophis
+
+# Initalise Acanthophis
+acanthophis-init .
 
 # the generate the fake dataset
 snakemake --snakefile Snakefile.generate-rawdata -j 8 --use-conda
 
 # and run the actual test dataset
-snakemake --snakefile Snakefile -j 8 --use-conda
+snakemake -j 8 --use-conda
 ```
 
 
-To initialise your own workflow, copy `./Snakefile` and `./config.yml` to a new
-directory. Establish a conda environment with some basic dependencies
-(`conda create -n workflow python snakemake mamba`). Install Acanthophis (`pip
-install -e $GITHUB_URL_OR_PATH_TO_CLONE_OF_ACANTOPHIS`). Then, collect your raw
-data, and create at least the two metadata files (modelled after
-`./data/rl2s.tsv` and `./data/samples.tsv`). Then, customise the config file
-and Snakefile to match your intended pipeline. The pipeline can then be run
-using something like `snakemake --snakefile Snakefile -j 8 --use-conda
---conda-frontend mamba`.
+To initialise your own workflow, run `acanthophis-init` in some new directory.
+Then, collect your raw data, and create at least the required metadata files
+and customise the config file to match your intended pipeline. You can then run
+the pipeline using something like any other snakemake file, for example
+`snakemake  -j 8 --use-conda`. Please follow [the documentation
+](../documentation.md), which outlines this process in much more detail.
 
-Some day soon, one will be able to do something like `acanthophis init` in some
-directory for commented boilerplate to be generated in place.
 
 A real deployment would obviously skip the creation of the fake dataset, so the
 below files are of no use to you:
