### 2024-01-09 kdm9/Acanthophis#14 @kdm9
Hi George, apologies, but I can't reproduce this in a fresh docker container. Could you please `pip install -U acanthophis` and confirm that this still happens in the latest version (which is now 0.3.0).
