### 2024-01-09 kdm9/Acanthophis#14 @kdm9
Also, do I guess correctly that you're on OSX? I really doubt that will cause issues, but I have never tested this on an OSX machine, so there is a minor chance there is some strange bug lurking on OSX.
