### 2024-01-11 c4465a5
add contribution/help to readme (#8)
