### 2024-01-11 kdm9/Acanthophis#9 @kdm9
This comment will track my progress through these minor changes, I'll update it as I go, and only close the issue when it's all done.

> Overall, the paper reads well, and I am convinced of your pipeline's value both in terms of need for the community, functionality and level of configurability. For the software paper to fully convey this, and to clarify certain points/features, I've provided the following comments below. They mostly relate to providing definitions, more details on the pipeline and use cases, and discussions of design choices and limitations.
>
> In two comments I refer to adding a possible 'Discussion' section to the paper. I'm not sure if this is something that JOSS does or not - if not, it's fine to address the comments directly on lines they refer to.
>
>     * For clarity the term 'Hologenomics' should be defined at the very beginning of the paper (currently 'hologenome' is defined only on Line 38).
>       E.g. in the paper you could tag it after the first sentence of the 'Summary' section:
>       '...joint discovery of both plant genetic variation and variation [...] of plant-associated microbiomes. We refer to this as hologenomics.'
>       Please also define this term in the repo README or `documentation.md`.
>
>     * Line 29: 'has attracted many users' do you have a metric for this (e.g. number of git clones, or PyPI downloads)? Otherwise, I would leave this statement out;
>       it's already obvious in my opinion that this pipeline is valuable, and you have studies that use it (see below).
>
>     * Lines 30-31: It seems that Ahrens et al. (2021) discusses using the pipeline only through referring to Murray et al. (2019), in which parts of an earlier version of this pipeline are discussed.
>       Perhaps state as such: 'this pipeline was first developed in Murray et al. (2019) for use in [x and y], and subsequently used
>       in Ahrens et al. (2021) for analysing species [x' and y']'.
>       Also, what are the improvements in the pipeline since Murray et al. (2019)? (I'm not saying there are none)
>       I believe you've added taxonomic identification since then - it would be good to state this.
>
>     * Relating to use cases: has Acanthophis been used already in a hologenomics context - i.e. identifying microbial species present on/near
>       focal plant species? If so, please give an example - it's not clear to me that Ahrens et al. (2021) or Murray et al. (2019) do this.
>       If not, can you discuss how for example Karasov et al. (2022) did 'hologenomics', did they build their own pipeline?
>
>     * Relating to hologenomics: I don't think your pipeline does host x microbe genetic association tests, perhaps add one sentence in a 'Discussion' section
>       about how a user could then seek to do this, after running Acanthophis?
>
>     * Line 29: Related to comment above, can you discuss Taxprofiler a little bit more - can Taxprofiler perform taxonomic assignment for plant
>       reads? If not, say that explicitly. Then, you could state that Acanthophis does both variant calling and taxonomic identification, and so
>       performs functionality of both Snakemake's variant calling pipeline and the nf-core Taxprofiler pipelines.
>       Does Acanthophis do anything that they don't? (perhaps genetic distances?) And (being devil's advocate) in what
>       cases is running Acanthophis better than running those two pipelines independently?
>       (I guess one obvious answer is analysing plant genomes; can Taxprofiler do hologenomics?)
>
>     * Line 36: Why can't your pipeline be run on long-read data (PacBio, ONT)? It is already valuable
>       for the pipeline to target short-read datasets, but it would be nice for the reader to understand
>       what prevents Acanthophis processing/being useful for long-read data.
>       This could be stated in a 'Discussion' section of the paper.
>
>     * Line 59: why should input reference genomes be uncompressed fasta files? I'm pretty sure that for at least
>       bwa-mem and minimap2, the fasta can be gzipped.
>
>     * Line 64: what does 'combines filter-passing alleles back into unique sites' mean? Two clauses back you say that before this filtering
>       you split multi-allelic variants, so I assume it's not related to that process?
>
>     * Lines 66-69: very nice paralellisation strategies!
>
>     * Lines 71-74: This sentence is v. hard to understand. I think it'd be clearer to say 'Acanthophis creates taxonomic profiles using any of Kraken2 [etc...],
>       or using other public sequence databases'. + Give examples of public sequence databases, and also whether the databases are automatically downloaded during execution.
>
>     * Line 77: For 'Stage 5: De novo Estimates of Genetic Dissimilarity', you provide means to compute genetic dissimilarity measures from the
>       reads themselves - i.e. alignment and variant-calling free, which is nice. Does Acanthopis also compute VCF-based dissimilarity measures?
>       If not, please state why not.
>
>     * Lines 81-84: please explicitly mention at least some of the tools/stages that MultiQC can run on, to get a sense of what the user can visualise and assess at the end.
>
>     * Perhaps after Line 84, state how independent the stages are: e.g. 'the user can perform all of variant calling, genetic distance estimation and taxonomic profiling isolation', if that's correct
