+++ tests/setup.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 mamba env update -f environment-setup.yml
-conda activate acanthophis-tests
+mamba activate acanthophis-tests
 rm -rf output/ tmp/
+mkdir -p output/ tmp/ rawdata/
+touch output/nobackup tmp/nobackup rawdata/nobackup
 pip install -e ../
 set -xeuo pipefail
 snakemake --snakefile Snakefile.generate-rawdata -j 4 --use-conda --conda-frontend mamba
