+++ tests/test.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
-conda activate acanthophis-tests
-set -xeuo pipefail
-python3 -m pip install ..
+set -euo pipefail
+mamba activate acanthophis-tests
+python3 -m pip uninstall --yes acanthophis
+python3 -m pip install -e ..
 acanthophis-init --yes
 mamba env update -f environment.yml
-conda activate acanthophis
+mamba activate acanthophis
 rm -fr output tmp
-snakemake -j 2 --use-conda --ri "${@}"
+set -x
+which snakemake
+snakemake --version
+snakemake -j $(nproc 2>/dev/null || echo 2)  --software-deployment-method conda apptainer --ri "${@}"
