+++ acanthophis/template/workflow/rules/envs/glnexus.yml
@@ -1,0 +1,7 @@
+channels:
+  - conda-forge
+  - bioconda
+dependencies:
+    - glnexus=1.4
+    - bcftools
+
