### 2024-01-16 kdm9/Acanthophis#12 @gbouras13
Hi @kdm9 ,

Yes, copying from the command line.

I'm getting back onto this now - will try and new pip install, will let you know how I go.

George
