### 2024-01-16 kdm9/Acanthophis#15 @gbouras13
[JOSS Review] Errors with Mac Installation
Re https://github.com/openjournals/joss-reviews/issues/6062

On my Macbook M1 I tried this:

```
git clone https://github.com/kdm9/Acanthophis.git
mamba env create -f environment-setup.yml -n acanthophis-demo
conda activate acanthophis-demo
pip install -U acanthophis # confirmed v0.3.0 was installed
cd Acanthophis/tests

acanthophis-init .

snakemake --snakefile Snakefile.generate-rawdata -j 8 --use-conda
```

This all ran no problems at all.

I then ran

```
snakemake snakemake -j 8 --use-conda --conda-frontend mamba
```

which errored:

```
Traceback (most recent call last):
  File "/Users/a1667917/mambaforge/envs/acanthophis-demo/lib/python3.11/site-packages/snakemake/cli.py", line 1886, in args_to_api
    dag_api = workflow_api.dag(
              ^^^^^^^^^^^^^^^^^
  File "/Users/a1667917/mambaforge/envs/acanthophis-demo/lib/python3.11/site-packages/snakemake/api.py", line 326, in dag
    return DAGApi(
           ^^^^^^^
  File "<string>", line 6, in __init__
  File "/Users/a1667917/mambaforge/envs/acanthophis-demo/lib/python3.11/site-packages/snakemake/api.py", line 425, in __post_init__
    self.workflow_api._workflow.dag_settings = self.dag_settings
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/a1667917/mambaforge/envs/acanthophis-demo/lib/python3.11/site-packages/snakemake/api.py", line 372, in _workflow
    workflow.include(
  File "/Users/a1667917/mambaforge/envs/acanthophis-demo/lib/python3.11/site-packages/snakemake/workflow.py", line 1369, in include
    exec(compile(code, snakefile.get_path_or_uri(), "exec"), self.globals)
  File "/Users/a1667917/Documents/Reviews/acanthophis/Acanthophis/tests/Snakefile", line 6, in <module>
AttributeError: module 'acanthophis' has no attribute 'populate_metadata'
```

George
