### 2024-01-16 kdm9/Acanthophis#14 @kdm9
@gbouras13 A thought: this could be due to poor interaction between installed pythons. Could you please check `which pip` (I think that works on osx, maybe `command -v pip`?). It should be a path under the active conda env, not e.g. /usr/bin/pip or some other dir, e.g. in homebrew.

This is an issue that would apply to all python packages, and is one of the rougher edges around python packaging.
