### 2024-01-16 kdm9/Acanthophis#15 @kdm9
Whie a different error, this is the same fundamental issue as #14, closing to avoid duplicating discussion
