### 2024-01-16 1dc2dc3
define hologenome
