### 2024-01-16 kdm9/Acanthophis#16 @gbouras13
Hi Kevin,

I got a `Illegal instruction (core dumped)` error with Freebayes v1.3.7.

I actually have run into issues with Freebayes v1.3.7 in a program of mine, pypolca, so I tried v1.3.6

```mamba install freebayes==1.3.6```

after I did this, acanthophis finished with no error.
