### 2024-01-16 fa614af
Update varcall.yml to fix freebayes <1.3.7
