+++ acanthophis/template/workflow/rules/envs/varcall.yml
@@ -4,5 +4,5 @@ channels:
 dependencies:
   - samtools
   - bcftools
-  - freebayes
+  - freebayes >=1.3.1,<1.3.7
 
