### 2024-01-16 kdm9/Acanthophis#16 @gbouras13
I've made a small commit and PR to fix this in `varcall.yaml` - this should solve the issue.

I'll finish the rest of my review tomorrow now.

George
