### 2024-01-16 kdm9/Acanthophis#16 @kdm9
yeah, ok, so that's independent from acanthophis then. I'm assuming you're using an old server that recent conda packages don't support?

I'm closing this, since it seems unrelated to acanthophis, but please reopen if there are further issues :)
