### 2024-01-16 kdm9/Acanthophis#16 @gbouras13
No @kdm9  I don't think that is an issue with my server (my server is very new) - the v1.3.7 version of `freebayes` on bioconda is the newest, so it will be installed by default for most people unless you set a dependency limit <1.3.7. It is an external issue to Acanthophis but it will otherwise still affect it.

George
