### 2024-01-16 ea52e68
GHA: stricter install versions
