+++ tests/environment-setup.yml
@@ -3,8 +3,8 @@ channels:
   - conda-forge
   - bioconda
 dependencies:
-  - python
-  - snakemake
+  - python>=3.11
+  - snakemake>=8
   - tree
   - pip
   - numpy
@@ -12,3 +12,4 @@ dependencies:
   - centrifuge-core
   - samtools
   - msprime
+  - curl
