+++ setup.cfg
@@ -22,13 +22,13 @@ classifiers =
 [options]
 packages = acanthophis
 include_package_data = True
-python_requires = >=3.7
+python_requires = >=3.11
 setup_requires =
     setuptools_scm
     setuptools_scm_git_archive
 install_requires =
     natsort
-    snakemake
+    snakemake>=8
 
 [bdist_wheel]
 universal=1
