### 2024-01-16 685254b
GHA: remove obsolte step
