### 2024-01-16 ae82d22
tests: fix snakemake sdm arg
