### 2024-01-16 cdf2b93
tests: always use python3 -m pip
