+++ tests/setup.sh
@@ -4,7 +4,7 @@ mamba activate acanthophis-tests
 rm -rf output/ tmp/
 mkdir -p output/ tmp/ rawdata/
 touch output/nobackup tmp/nobackup rawdata/nobackup
-pip install -e ../
+python3 -m pip install -e ../
 set -xeuo pipefail
 snakemake --snakefile Snakefile.generate-rawdata -j 4 --software-deployment-method conda
 tree rawdata
