### 2024-01-16 405e5a2
tests: use conda activate not mamba activate
