+++ tests/setup.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 mamba env update -f environment-setup.yml
-mamba activate acanthophis-tests
+conda activate acanthophis-tests
 rm -rf output/ tmp/
 mkdir -p output/ tmp/ rawdata/
 touch output/nobackup tmp/nobackup rawdata/nobackup
