### 2024-01-16 2a3d9f0
GHA: update setup-mamba action
