+++ .github/workflows/run-snakemake.yml
@@ -14,13 +14,13 @@ jobs:
       - uses: actions/checkout@v3
       
       - name: setup-mamba
-        uses: conda-incubator/setup-miniconda@v2
+        uses: conda-incubator/setup-miniconda@v3
         with:
+          python-version: 3.11
           mamba-version: "*"
           channels: bioconda,conda-forge
           channel-priority: true
           auto-update-conda: true
-            
       - name: Cache generated inputs
         id: cache-inputs
         uses: actions/cache@v3
