### 2024-01-16 3a01274
GHA: auto activate base initially
