+++ .github/workflows/run-snakemake.yml
@@ -21,6 +21,8 @@ jobs:
           channels: bioconda,conda-forge
           channel-priority: true
           auto-update-conda: true
+          auto-activate-base: true
+          activate-environment: ""
       - name: Cache generated inputs
         id: cache-inputs
         uses: actions/cache@v3
