### 2024-01-16 79b5fa1
GHA: use bash login shells
