+++ .github/workflows/run-snakemake.yml
@@ -21,8 +21,7 @@ jobs:
           channels: bioconda,conda-forge
           channel-priority: true
           auto-update-conda: true
-          auto-activate-base: true
-          activate-environment: ""
+          activate-environment: true
       - name: Cache generated inputs
         id: cache-inputs
         uses: actions/cache@v3
