+++ acanthophis/template/workflow/config.schema.yml
@@ -179,14 +179,6 @@ properties:
             snpeff:
               type: boolean
               default: false
-            deepvariant_model:
-              type: string
-              enum:
-                - WGS
-                - WES
-                - ONT_R104
-                - HYBRID_PACBIO_ILLUMINA
-                - PACBIO
   tool_settings:
     type: object
     properties:
@@ -233,6 +225,14 @@ properties:
         varcall:
           type: object
           properties:
+            deepvariant_model:
+              type: string
+              enum:
+                - WGS
+                - WES
+                - ONT_R104
+                - HYBRID_PACBIO_ILLUMINA
+                - PACBIO
             minmapq:
               type: object
               properties:
