### 2024-01-16 1d47404
tests: pin snakemake version
