+++ acanthophis/template/environment.yml
@@ -4,6 +4,6 @@ channels:
   - bioconda
 dependencies:
   - python>=3.11
-  - snakemake>=8
+  - snakemake=8.1.2
   - singularity
   - natsort
