### 2024-01-16 0d927f0
tests: pin snakemake 8.2, fixes version inconsistency
