+++ acanthophis/template/environment.yml
@@ -4,6 +4,7 @@ channels:
   - bioconda
 dependencies:
   - python>=3.11
-  - snakemake=8.1.2
+  - snakemake>=8.2
+  - snakemake-interface-executor-plugins>=8.2
   - singularity
   - natsort
