### 2024-01-16 83a3625
GHA: disable deep variant
