+++ tests/test.sh
@@ -8,4 +8,4 @@ mamba env update -f environment.yml
 conda activate acanthophis
 rm -fr output tmp
 set -x
-snakemake -j $(nproc 2>/dev/null || echo 2)  --software-deployment-method conda apptainer --ri "${@}"
+snakemake -j $(nproc 2>/dev/null || echo 2)  --software-deployment-method conda --ri "${@}"
