### 2024-01-16 540c554
better test script structure
