+++ acanthophis/template/config.yml
@@ -114,8 +114,8 @@ samplesets:
       callers:
         - mpileup
         - freebayes
-        # Deepvariant is disabled for Github actions tests, as it requires too much disk space
-        #- deepvariant
+        - deepvariant
+
       # Which short read aligners to use for variant calling? (can be
       # more/less/different to the align section above)
       aligners:
