+++ tests/.config_quick.yml
@@ -1,0 +1,231 @@
+#######################################################################
+#                             Data Paths                              #
+#######################################################################
+data_paths:
+  # Metadata locations.
+  metadata:
+    # This file maps the FASTQ files associated with a run of a library to a
+    # samplename, and to any other relevant metadata.
+    runlib2samp_file: "rawdata/rl2s.tsv"
+    # This is a shell glob pattern defining the text files containing lists of
+    # samplenames per sample set.
+    setfile_glob: "rawdata/samplesets/*.txt"
+
+  # Reference Genomes
+  references:
+    lambda:
+      fasta: "rawdata/reference/genome.fa"
+
+  # Taxon profiling databases
+  kraken:
+    Viral:
+      dir: "rawdata/kraken/Viral"
+      # If the database contains Bracken dbs, please uncomment the below and
+      # specify the bracken db sequence/kmer length to use
+      #bracken: 150
+  kaiju:
+    Viral:
+      nodes: "rawdata/kaiju/Viral/nodes.dmp"
+      fmi: "rawdata/kaiju/Viral/kaiju_db_viruses.fmi"
+  centrifuge:
+    lambda: "rawdata/centrifuge/lambda/lambda.1.cf"
+  # This directory should contain nodes.dmp and names.dmp from NCBI's taxdump.tar.gz
+  ncbi_taxonomy: "rawdata/ncbitax/"
+
+  # These are prefixes for output files, either temporary or persistent.
+  temp_prefix: "tmp/"
+  persistent_prefix: "output/"
+
+#######################################################################
+#                      Sample Set Configuration                       #
+#######################################################################
+#
+# This section is where we tell snakemake which files to generate for each set
+# of samples.  Samplesets are configured as files of sample names (see
+# setfile_glob above). 
+
+samplesets:
+
+  # `all_samples` is a inbuilt sample set, corresponding to all samples in the
+  # metadata file. If you only have one logical set of samples, you can use
+  # this as the sampleset name. If you have multiple sample sets, please
+  # duplicate this entire section for each sample set, and modify the settings
+  # accordingly.
+  all_samples:
+    
+    # Alignment of reads to a reference
+    align:
+      
+      # Which aligners should be used to map reads?
+      aligners:
+        - bwa
+        #- ngm
+      
+      # Against which references should we align?
+      references:
+        # Remember, each reference here must be defined in
+        # data_paths/references above.
+        - lambda
+
+      # Should we extract unmapped reads from each BAM?
+      unmapped_reads: true
+      # Should BAMs be kept longer than needed?
+      keep_bams: false 
+      # Generate samtools statistics?
+      stats: false
+      # Use qualimap to generate additional statistics?
+      qualimap: true
+
+    # Taxonomic classification
+    kraken:
+      dbs:
+        # Remember, each database here must be defined in data_paths/kraken
+        # above.
+        - Viral
+      # Output (un)-classified read fastqs from Kraken?
+      reads: false
+    #kaiju:
+    #  dbs:
+    #    # Remember, each database here must be defined in data_paths/kaiju
+    #    # above.
+    #    - Viral
+    #centrifuge:
+    #  dbs: []
+    #    # Centrifuge is disabled in this example as we use this in CI testing
+    #    # and centrifuge uses too much RAM, causing our tests to fail. To run
+    #    # centrifuge, supply an array/list of databases here as one does above
+    #    # for Kraken & Kaiju, for exampe:
+    #    #
+    #    # - lambda
+    
+    # Make a sample_R1/R2.fastq.gz file per sample?
+    persample_reads: false
+
+    # Run FastQC per input run/library?
+    fastqc: true
+
[…]
