+++ tests/README.md
@@ -2,15 +2,23 @@
 
 
 This directory contains a somewhat contrived but complete example workflow
-using the Acanthophis pipeline. It consists of two stages, a first that does a
+using the Acanthophis pipeline, primarily for continuous integration testing (to catch bugs early).
+It consists of two stages, a first that does a
 simplistic coalescent simulation of population variation, using the lambda
 phage genome as an ancestral genome (see `./Snakefile.generate-rawdata`). The
-second stage applies the read qc, aligment to ref, and variant calling
-pipelines from Acanthophis to call variants in this simulated dataset (see
-`./Snakefile` and `./config.yml`). 
+second stage applies the read QC, alignment to reference, and variant calling
+pipelines from Acanthophis to call variants in this simulated dataset.
 
+To run this test pipeline, one must:
 
-To run this pipeline, one must:
+### Quick mode: the core pipeline using the CI suite.
+
+Two scripts 
+```bash
+
+## Complete version
+
+Either `bash setup.sh && bash test.sh`, or:
 
 ```bash
 conda env create -f environment-setup.yml -n acanthophis-demo
@@ -29,17 +37,10 @@ snakemake -j 8 --use-conda
 ```
 
 
-To initialise your own workflow, run `acanthophis-init` in some new directory.
-Then, collect your raw data, and create at least the required metadata files
-and customise the config file to match your intended pipeline. You can then run
-the pipeline using something like any other snakemake file, for example
-`snakemake  -j 8 --use-conda`. Please follow [the documentation
+To initialise your own workflow, [follow the main
+documentation](../documentation.md). Briefly, run `acanthophis-init` in some
+new directory. Then, collect your raw data, and create at least the required
+metadata files and customise the config file to match your intended pipeline.
+You can then run the pipeline using something like any other snakemake file,
+for example `snakemake  -j 8 --use-conda`. Please follow [the documentation
 ](../documentation.md), which outlines this process in much more detail.
-
-
-A real deployment would obviously skip the creation of the fake dataset, so the
-below files are of no use to you:
-
-- `Snakefile.generate-rawdata`
-- `lambda/*`
-- `conda.yml`
