### 2024-01-16 df8a351
tests: re-add missing mkdir
