### 2024-01-18 kdm9/Acanthophis#14 @gbouras13
Hi @kdm9 ,

Can confirm that pip is in the right place

`/Users/a1667917/mambaforge/envs/acanthophis/bin/pip`

One thing - Have you considered putting acanthophis on bioconda? Then perhaps it might be slightly easier to install if this is an issue with pip vs conda.

George
