### 2024-01-18 #6062 @gbouras13
Acanthophis is a plant hologenomics pipeline implemented in Snakemake. It makes use of Snakemake best practices (separately envs for each rule, clearly defined rules etc) and is implemented via pip, creating separate conda environments for each rule automatically. It also has customisable profiles and resources usage settings allowing for ‘embarrassingly’ parallel computation which I imagine will be very useful in such datasets.

Overall the paper is clear, outlines the statements of need and explains what acanthophis does. All the constituent programs seem reasonable in achieving what is desired within acanthophis.

Software

After a few teething installation bugs were resolved, acanthophis ran extremely fast on the test dataset. I am still having Mac installation bugs, but I consider this to be not particularly relevant, as almost any realistic use case for acanthophis will be on a server or HPC environment due to the data size on the inputs.

The config file and docs are well commented and makes it clear for the settings that can be configured. The manifest input file is also well explained.

I have a few comments to make & improvements to suggest.

These 2 things I think are necessary before I recommend acanthophis:

1.	I think the documentation is a bit sparse. Some things I think should be fleshed out:
a.	Some explanation of outputs (at least at the subdirectory level). While they are clearly named in the outputs, it would be good to explicitly make it clear what outputs are where to a user.
b.	Some explanation of how/where to install the taxonomic profiling databases – at a minimum links to the websites where they can be downloaded (e.g. Kraken, Kaiju, Centrifuge).
2.	Acanthophis should be available on bioconda as well as PyPI. Given it requires conda/mamba anyway, it makes sense it have it available there. If you need a hand @kdm9 let me know

I also have one other suggestion that I believe would improve Acanthophis quite a bit, but not required (as it would probably be a fair amount of work to implement) – I would like to hear your thoughts @kdm9.

One aspect thing that I didn’t like much about installing and using Acanthophis was the requirement to copy rules and snakefiles. This isn’t a fatal flaw but is annoying and may cause confusion for some users (and also means there will be lots of copies of these files floating around potentially).

A way of solving this problem is to give Acanthophis a CLI interface – the user could modify and specify the config file and manifest and feed those in as the input.

Two options I know of to achieve this (n.B. I have collaborated with the author of Snaketool and used it myself) are Snaketool  (https://github.com/beardymcjohnface/Snaketool) and snk (https://github.com/Wytamma/snk).

This concludes my review. I've left a few boxes unticked for now.

George
