### 2024-01-18 kdm9/Acanthophis#14 @kdm9
OK, I am genuinely unable to reproduce this on any system I try. This is troubling and I want to fix this though, as this should not be any users first experience of a program. Is there any other details or debugging you can do? As far as I can tell, I ran your exact commands below, so it should have identical behaviour if the issue is as simple as a missing function or import.

About conda: yes, I have a draft package in my channel, but it's linux only, and is slightly out of date. I'll update it quickly now, and then you should be able to `mamba install -c kdm801 acanthophis=0.3.1`.

```
docker run -it --rm condaforge/miniforge3

# within docker:
root@368cfb25bd0a:/# mamba create -n someproject -c bioconda -c conda-forge python snakemake pip natsort
<<<<SNIP LOGO>>>>>
Looking for: ['python', 'snakemake', 'pip', 'natsort']

conda-forge/linux-64                                        Using cache
conda-forge/noarch                                          Using cache
bioconda/noarch                                      5.1MB @   3.7MB/s  1.4s
bioconda/linux-64                                    5.3MB @   3.7MB/s  1.5s
Transaction

  Prefix: /opt/conda/envs/someproject

  Updating specs:

   - python
   - snakemake
   - pip
   - natsort

  Package                                      Version  Build                Channel                   Size
─────────────────────────────────────────────────────────────────────────────────────────────────────────────
  Install:
─────────────────────────────────────────────────────────────────────────────────────────────────────────────

  + natsort                                      8.4.0  pyhd8ed1ab_0         conda-forge/noarch        37kB
  + python                                      3.12.1  hab00c5b_1_cpython   conda-forge/linux-64      32MB
  + snakemake                                    8.2.1  hdfd78af_0           bioconda/noarch            9kB
  + snakemake-interface-common                  1.15.1  pyhdfd78af_0         bioconda/noarch           18kB
  + snakemake-interface-executor-plugins         8.2.0  pyhdfd78af_0         bioconda/noarch           22kB
  + snakemake-interface-storage-plugins          3.0.0  pyhdfd78af_0         bioconda/noarch           18kB
  + snakemake-minimal                            8.2.1  pyhdfd78af_0         bioconda/noarch          207kB

<<<SKIPPED A LOT OF PKGS ABOVE>>>
  Summary:

  Install: 121 packages

  Total download: 104MB

Confirm changes: [Y/n]
 <<<<SKIP INSTALL LOG>>>>
Downloading and Extracting Packages

Preparing transaction: done
Verifying transaction: done
Executing transaction: done

To activate this environment, use

     $ mamba activate someproject

To deactivate an active environment, use

     $ mamba deactivate

(base) root@368cfb25bd0a:/Acanthophis# conda activate someproject
(someproject) root@368cfb25bd0a:/Acanthophis# pip install -e .
Obtaining file:///Acanthophis
  Installing build dependencies ... done
  Checking if build backend supports build_editable ... done
  Getting requirements to build editable ... done
  Preparing editable metadata (pyproject.toml) ... done
Requirement already satisfied: natsort in /opt/conda/envs/someproject/lib/python3.12/site-packages (from acanthophis==0.3.1.dev26) (8.4.0)
Requirement already satisfied: snakemake>=8 in /opt/conda/envs/someproject/lib/python3.12/site-packages (from acanthophis==0.3.1.dev26) (8.2.1)
<<<<<SKIP LOTS MORE "Requirement already satisfied:" >>>>
Building wheels for collected packages: acanthophis
  Building editable for acanthophis (pyproject.toml) ... done
  Created wheel for acanthophis: filename=acanthophis-0.3.1.dev26-0.editable-py2.py3-none-any.whl size=9818 sha256=06c58830d1490df1b2953ffa776e3862a8ed835b987d18a8cc6741af065045cb
  Stored in directory: /tmp/pip-ephem-wheel-cache.....
Successfully built acanthophis
Installing collected packages: acanthophis
Successfully installed acanthophis-0.3.1.dev26
WARNING: Running pip as the 'root' user can result in broken permissions and conflicting behaviour with the system package manager. It is recommended to use a virtual environment instead: https://pip.pypa.io/warnings/venv
(someproject) root@368cfb25bd0a:/Acanthophis# acanthophis-init some_project
cp -r /Acanthophis/acanthophis/template/ -> some_project? (WARNING: make sure you have git add'd all files as they will be overwritten)  [yN] y
(someproject) root@368cfb25bd0a:/Acanthophis#

```
