### 2024-01-18 kdm9/Acanthophis#18 @kdm9
Additional CLI entry point for running the whole workflow
Likely via snaketool. See https://github.com/openjournals/joss-reviews/issues/6062#issuecomment-1897579241
