### 2024-01-18 #6062 @kdm9
*this is a "live" comment I'll use to address George's feedback piece-by-piece. I'll remove this notice when I'm done*

> Acanthophis is a plant hologenomics pipeline implemented in Snakemake. It makes use of Snakemake best practices (separately envs for each rule, clearly defined rules etc) and is implemented via pip, creating separate conda environments for each rule automatically. It also has customisable profiles and resources usage settings allowing for ‘embarrassingly’ parallel computation which I imagine will be very useful in such datasets.
>
> Overall the paper is clear, outlines the statements of need and explains what acanthophis does. All the constituent programs seem reasonable in achieving what is desired within acanthophis.

Thanks George!

> Software
>
> After a few teething installation bugs were resolved, acanthophis ran extremely fast on the test dataset. I am still having Mac installation bugs, but I consider this to be not particularly relevant, as almost any realistic use case for acanthophis will be on a server or HPC environment due to the data size on the inputs.

I agree, but would still ideally like to solve those OSX bugs, as there is nothing obvious that should cause them to be unique to OSX, and so could effect other users. But I can't reproduce them on any system I've tried (including in fresh docker images), but that is all up-to-date linux. See https://github.com/kdm9/Acanthophis/issues/14#issuecomment-1899167421

> The config file and docs are well commented and makes it clear for the settings that can be configured. The manifest input file is also well explained.
>
> I have a few comments to make & improvements to suggest.
>
> These 2 things I think are necessary before I recommend acanthophis:
>
>     1. I think the documentation is a bit sparse. Some things I think should be fleshed out:
>        a.	Some explanation of outputs (at least at the subdirectory level). While they are clearly named in the outputs, it would be good to explicitly make it clear what outputs are where to a user.
>        b.	Some explanation of how/where to install the taxonomic profiling databases – at a minimum links to the websites where they can be downloaded (e.g. Kraken, Kaiju, Centrifuge).
>
>     2. Acanthophis should be available on bioconda as well as PyPI. Given it requires conda/mamba anyway, it makes sense it have it available there. If you need a hand @kdm9 let me know

I have a draft package, and was waiting before I cut a 1.0 release after any major peer review updates before shipping this in bioconda. But I will make sure this happens before the review is accepted.

>
> I also have one other suggestion that I believe would improve Acanthophis quite a bit, but not required (as it would probably be a fair amount of work to implement) – I would like to hear your thoughts @kdm9.
>
> One aspect thing that I didn’t like much about installing and using Acanthophis was the requirement to copy rules and snakefiles. This isn’t a fatal flaw but is annoying and may cause confusion for some users (and also means there will be lots of copies of these files floating around potentially).
>
> A way of solving this problem is to give Acanthophis a CLI interface – the user could modify and specify the config file and manifest and feed those in as the input.
>
> Two options I know of to achieve this (n.B. I have collaborated with the author of Snaketool and used it myself) are Snaketool (https://github.com/beardymcjohnface/Snaketool) and snk (https://github.com/Wytamma/snk).

I've actually gone back and forth on this a few times. In a much earlier version, did have Acanthophis set up with a CLI entry point, in a similar fashion to how snaketool operates (though before snaketool itself existed). But, it became rather inflexible and harder to maintain, and the core rule files were hidden from the end user. Feedback from several early Acanthophis users was that they preferred having a snakemake pipeline plonked in front of them, that they could run and modify as though they were the original authors.

I'll certainly keep Acanthophis operating in its current form, but I'll investigate setting it up so that once pip/conda installed, there's an additional snaketool entrypoint that allows running an unmodified version of the pipeline from `$PREFIX/lib/python3/site-packages/acanthophis/...` as a CLI, for users who preference a clean directory and a single CLI command over flexibility. This might be in a subsequent release though, and I've added a github issue to remind me (https://github.com/kdm9/Acanthophis/issues/18)

> This concludes my review. I've left a few boxes unticked for now.
>
> George

Thanks again for an extremely helpful and thorough review George, very much appreciated.

Best,
Kevin
