### 2024-01-22 kdm9/Acanthophis#19 @abhishektiwari
Included Dockerfile does not build
Running `docker build -t acanthophis .` from root of repository and throws following error. It seems Dockerfile is outdated.

```
16.77 INFO: pip is looking at multiple versions of acanthophis to determine which version is compatible with other requirements. This could take a while.
16.77 ERROR: Ignored the following versions that require a different python version: 8.0.0 Requires-Python >=3.11; 8.0.1 Requires-Python >=3.11; 8.1.0 Requires-Python >=3.11; 8.1.1 Requires-Python >=3.11; 8.1.2 Requires-Python >=3.11; 8.1.3 Requires-Python >=3.11; 8.2.0 Requires-Python >=3.11; 8.2.1 Requires-Python >=3.11; 8.2.2 Requires-Python >=3.11; 8.2.3 Requires-Python >=3.11
16.77 ERROR: Could not find a version that satisfies the requirement snakemake>=8 (from acanthophis) (from versions: 1.0, 1.0.1, 1.0.2, 1.1, 1.1.1, 1.1.2, 1.1.3, 1.1.4, 1.2, 1.2.1, 1.2.3, 2.0, 2.1, 2.1.1, 2.2, 2.2.1, 2.2.2, 2.3, 2.4, 2.4.1, 2.4.2, 2.4.3, 2.4.4, 2.4.5, 2.4.6, 2.4.7, 2.4.7.1, 2.4.8, 2.4.9, 2.5, 2.5.1, 2.5.2, 2.5.2.1, 2.5.2.2, 3.0b1, 3.0, 3.1, 3.1.1, 3.2, 3.2.1, 3.2.2, 3.3, 3.4, 3.4.1, 3.4.2, 3.5.0, 3.5.1, 3.5.2, 3.5.3, 3.5.4, 3.5.5, 3.6.0, 3.6.1, 3.7.0, 3.7.1, 3.8.0, 3.8.1, 3.8.2, 3.9.0, 3.9.1, 3.10.0, 3.10.1, 3.10.2, 3.11.0, 3.11.1, 3.11.2, 3.12.0, 3.13.0, 3.13.1, 3.13.2, 3.13.3, 4.0.0, 4.1.0, 4.2.0, 4.3.0, 4.3.1, 4.4.0, 4.5.0, 4.5.1, 4.6.0, 4.7.0, 4.8.0, 4.8.1, 5.0.0, 5.1.1, 5.1.2, 5.1.3, 5.1.4, 5.1.5, 5.2.0, 5.2.1, 5.2.2, 5.2.3, 5.2.4, 5.3.0, 5.3.1, 5.4.0, 5.4.1, 5.4.2, 5.4.3, 5.4.4, 5.4.5, 5.5.0, 5.5.1, 5.5.2, 5.5.3, 5.5.4, 5.6.0, 5.7.0, 5.7.1, 5.7.4, 5.8.0, 5.8.1, 5.8.2, 5.9.0, 5.9.1, 5.10.0, 5.11.0, 5.11.1, 5.11.2, 5.12.0, 5.12.1, 5.12.2, 5.12.3, 5.13.0, 5.14.0, 5.15.0, 5.16.0, 5.17.0, 5.18.0, 5.18.1, 5.19.0, 5.19.1, 5.19.2, 5.19.3, 5.20.0, 5.20.1, 5.21.0, 5.22.0, 5.22.1, 5.23.0, 5.24.0, 5.24.1, 5.24.2, 5.25.0, 5.26.0, 5.26.1, 5.27.2, 5.27.3, 5.27.4, 5.28.0, 5.29.0, 5.30.0, 5.30.1, 5.30.2, 5.31.0, 5.31.1, 5.32.0, 5.32.1, 5.32.2, 6.0.0, 6.0.1, 6.0.2, 6.0.3, 6.0.4, 6.0.5, 6.1.0, 6.1.1, 6.1.2, 6.2.0, 6.2.1, 6.3.0, 6.4.0, 6.4.1, 6.5.0, 6.5.1, 6.5.2, 6.5.3, 6.5.5, 6.6.0, 6.6.1, 6.7.0, 6.8.0, 6.8.1, 6.8.2, 6.9.0, 6.9.1, 6.10.0, 6.11.0, 6.11.1, 6.12.0, 6.12.1, 6.12.2, 6.12.3, 6.13.0, 6.13.1, 6.14.0, 6.15.0, 6.15.1, 6.15.2, 6.15.3, 6.15.4, 6.15.5, 7.0.0, 7.0.1, 7.0.2, 7.0.3, 7.0.4, 7.1.0, 7.1.1, 7.2.0, 7.2.1, 7.3.0, 7.3.1, 7.3.2, 7.3.3, 7.3.4, 7.3.5, 7.3.6, 7.3.7, 7.3.8, 7.4.0, 7.5.0, 7.6.0, 7.6.1, 7.6.2, 7.7.0, 7.8.0, 7.8.1, 7.8.2, 7.8.3, 7.8.5, 7.9.0, 7.10.0, 7.11.0, 7.12.0, 7.12.1, 7.13.0, 7.14.0, 7.14.1, 7.14.2, 7.15.0, 7.15.1, 7.15.2, 7.16.0, 7.16.1, 7.16.2, 7.17.0, 7.17.1, 7.18.0, 7.18.1, 7.18.2, 7.19.0, 7.19.1, 7.20.0, 7.21.0, 7.22.0, 7.23.0, 7.23.1, 7.24.0, 7.24.1, 7.24.2, 7.25.0, 7.25.1, 7.25.2, 7.25.3, 7.25.4, 7.26.0, 7.28.0, 7.28.1, 7.28.2, 7.28.3, 7.29.0, 7.30.0, 7.30.1, 7.30.2, 7.31.0, 7.31.1, 7.32.0, 7.32.1, 7.32.2, 7.32.3, 7.32.4)
16.77 ERROR: No matching distribution found for snakemake>=8
------
Dockerfile:5
--------------------
   3 |     WORKDIR /usr/local/src/acanthophis
   4 |     COPY . .
   5 | >>> RUN conda config --add channels bioconda &&  mamba install --yes --all snakemake && pip install -e .
   6 |
--------------------
ERROR: failed to solve: process "/bin/sh -c conda config --add channels bioconda &&  mamba install --yes --all snakemake && pip install -e ." did not complete successfully: exit code: 1
```

For [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6062)
