### 2024-01-22 kdm9/Acanthophis#7 @abhishektiwari
Using `environment-setup.yml` I could not run test as well.

SSH to container,

```
docker run -it --rm condaforge/miniforge3
```

Then run

```
git clone https://github.com/kdm9/Acanthophis.git
```

```
(acanthophis-tests) root@5110a2faf4b9:/Acanthophis/tests# mamba env update -f environment-setup.yml
conda-forge/linux-aarch64                                   Using cache
conda-forge/noarch                                          Using cache
bioconda/linux-aarch64                                      Using cache
bioconda/noarch                                             Using cache

Looking for: ["python[version='>=3.11']", "snakemake[version='>=8.2']", "snakemake-interface-executor-plugins[version='>=8.2']", 'tree', 'pip', 'numpy', 'wgsim', 'centrifuge-core', 'samtools', 'msprime', 'curl']

Could not solve for environment specs
The following packages are incompatible
├─ centrifuge-core does not exist (perhaps a typo or a missing channel);
├─ msprime does not exist (perhaps a typo or a missing channel);
├─ samtools does not exist (perhaps a typo or a missing channel);
└─ wgsim does not exist (perhaps a typo or a missing channel).
```

I tried other methods to install `msprime`  but it does not work,

```
(acanthophis-tests) root@5110a2faf4b9:/Acanthophis/tests# conda install -c conda-forge msprime
Channels:
 - conda-forge
 - bioconda
Platform: linux-aarch64
Solving environment: failed

PackagesNotFoundError: The following packages are not available from current channels:

  - msprime

```
