### 2024-01-22 #6062 @abhishektiwari
## Installation
Installation instructions are inconsistent and create a lot of friction and confusion. Case in point [installation using mamba](https://github.com/kdm9/Acanthophis/blob/main/documentation.md#tldr) (*works using `condaforge/miniforge3` docker image*) vs. [installation using pip](https://github.com/kdm9/Acanthophis/blob/main/documentation.md#acanthophis-setup) (*does not work*). Some places [installing from source](https://github.com/kdm9/Acanthophis/blob/df8a3514438c9c728b75eda335fb2dfdc16bc54f/tests/test.sh#L7). After bit of struggle I was able to install the package, got it running, and validate documented functionalities.

## Documentation and Examples
Authors should consider including example input files and cleaning up documentation. Also, please provide a working Dockerfile. Currently included `Dockerfile` [fails to build](https://github.com/kdm9/Acanthophis/issues/19). I was [unable to run](https://github.com/kdm9/Acanthophis/issues/7#issuecomment-1902901718) example scripts included in `tests` folder.

## Suggestions:
1.  In long run, it will better if authors can wrap commands exposed by Acanthophis using [click](https://click.palletsprojects.com/en/8.1.x/) as a CLI which will be a lot easier to distributed and install. And also allow to update the pipeline configuration from CLI (as suggested by other reviewer)

2. Code quality and structure can be improved a lot. For instance, authors should consider a clear separation between templating and Python code otherwise project will be hard to maintain ([see a file here](https://github.com/kdm9/Acanthophis/blob/main/acanthophis/template/workflow/rules/base.rules)).

2. No automated testing to catch regression and bugs (`tests` is misleading it should be renamed as `example`). Suggest to add some minimal automated tests (ideally unit but even integration should suffice) but adding tests will require template and code decoupling as suggesting above.
