### 2024-01-22 a4c0b07
use miniforge3 not obsolte mambaforge docker img
