+++ Dockerfile
@@ -1,5 +1,9 @@
-FROM condaforge/mambaforge:latest
-
+FROM condaforge/miniforge3:latest
+RUN conda config --add channels defaults && \
+    conda config --add channels bioconda && \
+    conda config --add channels conda-forge && \
+    conda config --set channel_priority strict && \
+    mamba install snakemake=8 natsort python>=3.11
 WORKDIR /usr/local/src/acanthophis
 COPY . .
-RUN conda config --add channels bioconda &&  mamba install --yes --all snakemake && pip install -e .
+RUN pip install -e .
