### 2024-01-22 #6062 @kdm9
Thanks for the review.

> ## Installation
>
> Installation instructions are inconsistent and create a lot of friction and confusion. Case in point [installation using mamba](https://github.com/kdm9/Acanthophis/blob/main/documentation.md#tldr) (_works using `condaforge/miniforge3` docker image_) vs. [installation using pip](https://github.com/kdm9/Acanthophis/blob/main/documentation.md#acanthophis-setup) (_does not work_). Some places [installing from source](https://github.com/kdm9/Acanthophis/blob/df8a3514438c9c728b75eda335fb2dfdc16bc54f/tests/test.sh#L7). After bit of struggle I was able to install the package, got it running, and validate documented functionalities.

I have made it clearer in the docs that it is easiest if one uses conda/mamba to install the base environment, and pip to install acanthophis (until it enters bioconda, see discussion with George above).

> ## Documentation and Examples
>
> Authors should consider including example input files and cleaning up documentation.

Example input files **are** included. More specifically, I include a workflow to generate a test dataset of some GB under `tests/`. Because this is a simulation, there is no point in distributing the data itself, when any user who needs it can trivially create it with the provided workflow. This is in fact how the unit tests work.

> Also, please provide a working Dockerfile. Currently included `Dockerfile` [fails to build](https://github.com/kdm9/Acanthophis/issues/19). I was [unable to run](https://github.com/kdm9/Acanthophis/issues/7#issuecomment-1902901718) example scripts included in `tests` folder.

Dockerfile is fixed, thanks. It used the recently obsoleted mambaforge, which I've replaced with miniforge3.

> ## Suggestions:
>
>     1. In long run, it will better if authors can wrap commands exposed by Acanthophis using [click](https://click.palletsprojects.com/en/8.1.x/) as a CLI which will be a lot easier to distribute and install. And also allow to update the pipeline configuration from CLI (as suggested by other reviewer)

As said in the other review comments -- this will happen, but is not a priority as when surveyed, the early users specifically asked for the opposite. As tracked in https://github.com/kdm9/Acanthophis/issues/18 I will add this using `snaketool` in a future version.

>     2. Code quality and structure can be improved a lot. For instance, authors should consider a clear separation between templating and Python code otherwise project will be hard to maintain ([see a file here](https://github.com/kdm9/Acanthophis/blob/main/acanthophis/template/workflow/rules/base.rules)).

While there are a couple of utility function there that could in theory be refactored out, these functions mostly require access to workflow-global variables, and therefore are best placed within workflow code. I have separated them from the main rules of the workflow (note that there are no "real" rules in the base file), and that the templating code is completely separate (under `acanthophis/__init__.py` and `acanthophis/cmd.py`).

>     3. No automated testing to catch regression and bugs (`tests` is misleading it should be renamed as `example`). Suggest to add some minimal automated tests (ideally unit but even integration should suffice) but adding tests will require template and code decoupling as suggesting above.

Sorry, but that's simply untrue: the test directory is a simulated complete workflow that is run on every commit [via github actions](https://github.com/kdm9/Acanthophis/actions).  Those tests simulate a population sequencing workflow and run the entire acanthophis workflow on those files, and directly tests the initial example configuration generated by `acathophis init`.
