### 2024-01-23 4c1f911
tests: stop pinning snakemaek version
