+++ tests/environment-setup.yml
@@ -4,8 +4,7 @@ channels:
   - bioconda
 dependencies:
   - python>=3.11
-  - snakemake>=8.2
-  - snakemake-interface-executor-plugins>=8.2
+  - snakemake=8
   - tree
   - pip
   - numpy
