### 2024-01-23 6f8133f
docs: add linux only supported note to paper & readme
