### 2024-01-23 kdm9/Acanthophis#14 @kdm9
I just tried to reproduce this on a colleague's OSX laptop, and couldn't reproduce this there either. But, nor could I install several of the dependencies. I've added a note to the JOSS paper and to the readme that Acanthophis is only supported on Linux, because the underlying tools seem uninstallable on OSX, and therefore it's not exactly useful.
