### 2024-01-23 kdm9/Acanthophis#9 @kdm9
I've peppered some rephrasing of the above discussion throughout the paper draft, which hopefully answers a reader if they have the same questions as @bricoletc (& likewise other reviewers).
