### 2024-01-23 2c573ef
paper: minor tweaks incorporating peer review discussion
