### 2024-01-23 kdm9/Acanthophis#9 @kdm9
@bricoletc I think I've addressed everything here, so I'm closing this issue. I'll regenerate the JOSS pdf in a little bit, and ping you in the main review thread when I'm satisfied everything is addressed. For now I'm closing this issue, reopen it if you see anything outstanding.
