### 2024-01-23 #6062 @kdm9
@marcosvital, I think I've satisfactorily addressed nearly all concerns of the reviewers (thanks to you all!)

Outstanding issues:

- [ ] A snaketool CLI, which as discussed above is a nice-to-have for users that prefer a single CLI invocation to the flexibility of a deployed workflow. This will happen in a future release, as tracked at https://github.com/kdm9/Acanthophis/issues/18
- [ ] A conda package. A draft is available from  [my personal conda channel](https://anaconda.org/kdm801/acanthophis) and I have a [opened PR to bioconda](https://github.com/bioconda/bioconda-recipes/pull/45376) that adds the latest released acanthophis version.

I'd strongly prefer that Acanthophis is accepted despite the first point above, as this will take time I am not able to immediately commit. The conda package should happen reasonably quickly.

There were some issues with installation on OSX, mainly poor interactions between conda & pip, and some dependencies not packaged for OSX. I now formally note that while in theory a cross-platform package, there is little point to Acanthophis on OSX as a), it's for computations larger that what a Macbook can handle, and b), most of its dependencies are either not available in conda for OSX, or explicitly do not work on OSX. One can still install Acanthophis on OSX using pip, provided snakemake is installable, and could use the pipeline without conda, but this is not officially supported. Hopefully this is now clear to all users.
