### 2024-01-23 0c235db
Remove hard python >= 3.11 requirement
