+++ setup.cfg
@@ -22,7 +22,6 @@ classifiers =
 [options]
 packages = acanthophis
 include_package_data = True
-python_requires = >=3.11
 setup_requires =
     setuptools_scm
     setuptools_scm_git_archive
