### 2024-01-24 kdm9/Acanthophis#9 @bricoletc
Looks good to me! I'll re-read the updated JOSS pdf when I see it on the main review thread :))
