### 2024-01-24 kdm9/Acanthophis#7 @bricoletc
@kdm9 , I've tried running the steps here: https://github.com/kdm9/Acanthophis/tree/main/tests
At the step `snakemake --snakefile Snakefile.generate-rawdata -j 10 --software-deployment-method conda` when running `bash setup.sh`, I get the following error:
```sh
Error in rule ncbitax:
    jobid: 76
    output: rawdata/ncbitax
    log: rawdata/ncbitax/get.log (check log file(s) for error details)
    shell:
        (mkdir -p rawdata/ncbitax && cd rawdata/ncbitax && wget -c ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz && tar -zxvf taxdump.tar.gz) &>rawdata/ncbitax/get.log
        (one of the commands exited with non-zero exit code; note that snakemake uses bash strict mode!)
```
Here is the corresponding log file from that rule:
[get.log](https://github.com/kdm9/Acanthophis/files/14040406/get.log)

Also, I've proposed pointing to the tests in your README in #20 :))
