### 2024-01-24 8c88107
tests: no unbound variable without --quick
