### 2024-01-24 kdm9/Acanthophis#7 @kdm9
@bricoletc OK, that's super strange, the log file you attach seems to suggest a failed download (see how it continues past 100% of file downloaded in the progress indicator),  resulting in an invalid tar file. I just tried that URL download manually myself and it worked fine.

Any chance you could try `wget ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz` yourself?   I struggle to see how anything Acanthophis does could change what file gets downloaded, so perhaps there was an intermittent error/update in progress on NCBI's end. Out of an abundance of caution, I've changed the old `ftp://` link for `https://`, in case there is there is some odd interaction with FTP. You could re-try with the latest commit.

Many thanks for the readme update b.t.w.! :)

best,
K
