### 2024-01-24 658cbff
tests: automatically use apptainer when not in --quick mode
