### 2024-01-28 kdm9/Acanthophis#7 @bricoletc
@kdm9 welcome for the update!
Can confirm it must've been a temporary issue on the NCBI side, now works.

However I still get the same error, on acanthophis version 0.3.3.dev0 installed from pypi as per `test.sh`:
```sh
❯ snakemake -j $(nproc 2>/dev/null || echo 2)  --software-deployment-method conda --ri
Traceback (most recent call last):
  File "/home/adminbrice/Softs/miniforge3/envs/acanthophis/lib/python3.12/site-packages/snakemake/cli.py", line 1893, in args_to_api
    dag_api = workflow_api.dag(
              ^^^^^^^^^^^^^^^^^
  File "/home/adminbrice/Softs/miniforge3/envs/acanthophis/lib/python3.12/site-packages/snakemake/api.py", line 326, in dag
    return DAGApi(
           ^^^^^^^
  File "<string>", line 6, in __init__
  File "/home/adminbrice/Softs/miniforge3/envs/acanthophis/lib/python3.12/site-packages/snakemake/api.py", line 436, in __post_init__
    self.workflow_api._workflow.dag_settings = self.dag_settings
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/adminbrice/Softs/miniforge3/envs/acanthophis/lib/python3.12/site-packages/snakemake/api.py", line 383, in _workflow
    workflow.include(
  File "/home/adminbrice/Softs/miniforge3/envs/acanthophis/lib/python3.12/site-packages/snakemake/workflow.py", line 1379, in include
    exec(compile(code, snakefile.get_path_or_uri(), "exec"), self.globals)
  File "/home/adminbrice/Desktop/research/writing/reviewing_exernal/JOSS/Acanthophis/test_tool/Acanthophis/tests/Snakefile", line 6, in <module>
AttributeError: module 'acanthophis' has no attribute 'populate_metadata'
```
