### 2024-01-28 3ef9746
nuke all outputs before tests
