+++ tests/test.sh
@@ -1,22 +1,21 @@
 #!/bin/bash
-set -euo pipefail
 
 # Reinstall acanthophis from git
 conda activate acanthophis-tests
+set -eo pipefail
 python3 -m pip uninstall --yes acanthophis
 python3 -m pip install -e ..
 
 # re-init after reinstall
+rm -rf output/ tmp/ workflow config.yml Snakefile environment.yml
+mkdir -p output/ tmp/
+touch output/nobackup tmp/nobackup
 acanthophis-init --yes
 
 # recreate the conda env
 mamba env update -f environment.yml
 conda activate acanthophis
 
-# Clear outputs
-rm -fr output tmp
-mkdir -p output/ tmp/
-touch output/nobackup tmp/nobackup
 
 # If quick, overwrite config with cut-back version
 SDM=(conda)
