### 2024-01-28 kdm9/Acanthophis#7 @kdm9
@bricoletc thanks once again for helping debug this, but the only cause I can think of is an outdated copy of the snakefile in ./tests/. As you can see in the source, the [Snakefile](https://github.com/kdm9/Acanthophis/blob/main/acanthophis/template/workflow/Snakefile#L3) directly calls `populate_metadata()`, which is defined in [rules/base.rules](https://github.com/kdm9/Acanthophis/blob/658cbffd2c947709b4e57ca705ba7109e49f4c57/acanthophis/template/workflow/rules/base.rules#L48).

So, it's certainly true that `module 'acanthophis' has no attribute 'populate_metadata'`, but nor does it need to, since [at least 2022](https://github.com/kdm9/Acanthophis/commit/08c3e2b278bc591e8e6a67e2feb06314b333d02d#diff-065e77b19e7610afb1b85de65b0ecdbf52158e4aa04463b5ff1939104c840232L2-R4). But, if I remember correctly, when you first installed Acanthophis the last released (i.e. non-release candidate) version was from before this change. `acanthophis-init` by default does not overwrite files, so if you  generated that snakefile way back when you first installed Acanthophis, you might be stuck with an old version.

This is suboptimal, so I've just added code to `tests/test.sh` to nuke all acanthophis code before running `acanthophis-init`, so that it generates them afresh. This then matches what happens both in CI and in my local precommit hook tests, which happen in ephemeral docker containers.

So, can I please ask one last debugging check of you? Please check if the `Snakefile` in the tests directory matches [HEAD](https://github.com/kdm9/Acanthophis/blob/main/acanthophis/template/workflow/Snakefile) (e.g., from the root of the cloned repo: `diff tests/Snakefile acanthophis/template/workflow/Snakefile`). If there is a difference, please pull the latest commit in main with the updated `test.sh`, and then re-run.

Thanks a million for your persistence!

best,
K
