### 2024-02-01 kdm9/Acanthophis#9 @bricoletc
@kdm9 had a look at your updated pdf built [here](https://github.com/openjournals/joss-reviews/issues/6062#issuecomment-1919149593) and have a few comments:

(If I'm doing this too early, and you were going to do these anyway, apologies ;) )

- Term 'hologenomics' needs defining early on, e.g. in 'Summary' section
- `estimating genetic distances *de novo*`; I would maybe say `estimating genetic distances *de novo* (from the reads directly)`
- `This way, one can configure the analyses to be run (most can be disabled if not needed)` I would maybe say `This way, one can configure the analyses to be run (most stages described below can be disabled if not needed)`
- Regarding Taxprofiler, `most host-aware microbiome analysis pipelines do not allow for host genotyping and/or assume an animal host (e.g. Taxprofiler` Taxprofiler doesn't do any genotyping, does it? So can remove `host` here? Out of curiosity, does Acanthophis try and separately genotype the host and host-associated microbe(s)?

All my other comments are well addressed :))
