### 2024-02-01 kdm9/Acanthophis#9 @bricoletc
Woops!

(Btw, I saw @gbouras13's comment regarding Snaketool-like CLI functionality, and as a user I would definitely prefer to run the Snakemake workflow myself, so just saying I agree it's not strictly necessary for this paper)

(Going away now)
