### 2024-02-01 kdm9/Acanthophis#9 @kdm9
> @kdm9 had a look at your updated pdf built [here](https://github.com/openjournals/joss-reviews/issues/6062#issuecomment-1919149593) and have a few comments:
>
> (If I'm doing this too early, and you were going to do these anyway, apologies ;) )
>
>     * Term 'hologenomics' needs defining early on, e.g. in 'Summary' section

Done. it was sorta already defined, but just not mentioned by name, so hopefully clearer now.

>     * `estimating genetic distances *de novo*`; I would maybe say `estimating genetic distances *de novo* (i.e. from the reads directly)`

Clarified

>     * `This way, one can configure the analyses to be run (most can be disabled if not needed)` I would maybe say `This way, one can configure the analyses to be run (most of the stages described below can be turned on or off by the user)`

Fixed

>     * Regarding Taxprofiler, `most host-aware microbiome analysis pipelines do not allow for host genotyping and/or assume an animal host (e.g. Taxprofiler` Taxprofiler doesn't do any genotyping, does it? So can remove the word `host` in that sentence?

Done

> Out of curiosity, does Acanthophis try and separately genotype the host and host-associated microbe(s)?

So the answer here is slightly complex, but yes, it can. You can align reads & call variants against any reference genome. For example, on one project where we have a single pathogen of interest I have jointly called variants from wild infected samples against a concatenated host + pathogen genome. You could also call against some pangenome panel of bacterial pathogens or really whatever fasta file you can construct and index for BWA et al...

The interpretation & performance gets somewhat complex when you start e.g. making full flattened pangenomes of some bacterial genus (as I need not tell you!! :smile: ), but in terms of the tooling, you can certainly run those commands with Acanthophis, even if the result might not be what you expect.

Building something like pandora or gramtools in would be super cool, but I haven't needed to yet and nobody has asked me to, so it's on the infinite todo list for now.

> All my other comments are well addressed :))

Brilliant, thanks again.
