+++ paper/paper.bib
@@ -1,512 +1,579 @@
-@online{ahrens21_genomicconstraints,
-  title = {Genomic Constraints to Drought Adaptation},
-  author = {Ahrens, Collin W. and Murray, Kevin D. and Mazanec, Richard A. and Ferguson, Scott and Bragg, Jason and Jones, Ashley and Tissue, David T. and Byrne, Margaret and Borevitz, Justin O. and Rymer, Paul D.},
-  date = {2021-08-08},
-  eprinttype = {bioRxiv},
-  eprintclass = {New Results},
-  pages = {2021.08.07.455511},
-  doi = {10.1101/2021.08.07.455511},
-  url = {https://www.biorxiv.org/content/10.1101/2021.08.07.455511v1},
-  urldate = {2023-10-09},
-  abstract = {Global shifts in climate and precipitation patterns are altering the diversity and structure of forests. The ability for species to adapt is especially difficult for long lived foundation species with unknown genetic and trait diversity. We harnessed genomic, physiological, and climate data to determine adaptation constraints. We used denovo assembly and 6.5 million genomic variants with drought related traits from 432 individuals sourced from across the complete range of the foundation tree species Corymbia calophylla. We found genomic variants determining traits predominantly in gene regulatory regions. The ability for populations to adapt was limited by within population genetic diversity associated with traits, and epistatic interactions within traits and pleiotropic interactions among traits. Nevertheless, we could accurately predict adaptive traits using genomic and climate data to guide reforestation. Our study indicated that some populations may contain variation sufficient for the species to adapt to the effects of drought, while other populations will need increased variability from those sources.},
-  langid = {english},
-  pubstate = {preprint},
-  file = {/home/kevin/work/bibliography/zotdir/storage/UM62FW75/Ahrens et al. - 2021 - Genomic constraints to drought adaptation.pdf}
-}
 
-@article{beber23_taxpastataxonomic,
-  title = {{{TAXPASTA}}: {{TAXonomic Profile Aggregation}} and {{STAndardisation}}},
-  shorttitle = {{{TAXPASTA}}},
-  author = {Beber, Moritz E. and Borry, Maxime and Stamouli, Sofia and Yates, James A. Fellows},
-  date = {2023-07-11},
-  journaltitle = {Journal of Open Source Software},
-  shortjournal = {J. Open Source Softw.},
-  volume = {8},
-  number = {87},
-  pages = {5627},
-  issn = {2475-9066},
-  doi = {10.21105/joss.05627},
-  url = {https://joss.theoj.org/papers/10.21105/joss.05627},
-  urldate = {2023-10-04},
-  abstract = {Beber et al., (2023). TAXPASTA: TAXonomic Profile Aggregation and STAndardisation. Journal of Open Source Software, 8(87), 5627, https://doi.org/10.21105/joss.05627},
-  langid = {english},
-  file = {/home/kevin/work/bibliography/pdfs/beber23_taxpastataxonomic__journal_of_open_source_software__taxpasta.pdf}
-}
-
-@article{buchfink15_fastsensitive,
-  title = {Fast and Sensitive Protein Alignment Using {{DIAMOND}}},
-  author = {Buchfink, Benjamin and Xie, Chao and Huson, Daniel H.},
-  date = {2015-01},
-  journaltitle = {Nature Methods},
-  shortjournal = {Nat Meth},
-  volume = {12},
-  number = {1},
-  pages = {59--60},
-  issn = {1548-7091},
-  doi = {10.1038/nmeth.3176},
-  url = {http://www.nature.com/nmeth/journal/v12/n1/full/nmeth.3176.html},
-  urldate = {2016-07-13},
[…]
