### 2024-02-08 #6062 @kdm9
@marcosvital I'm just finalizing the edits now, based on final feedback from coauthors. I'll post here tagging you and all reviewers when the hopefully-final copy is ready (later today hopefully).

By pointing to specific modifications, do you mean a diffed paper version, or changes to the underlying code? Mainly the changes have been small but important bug fixes, and cleaning the paper or documentation of confusing, out of date, or otherwise wrong statements. I can provide more details if you need them, please just ask.
