+++ paper/paper.md
@@ -25,7 +25,7 @@ authors:
     affiliation: "2,3"
     corresponding: true
 affiliations:
- - name: Max Planck Institute for Biology, Tübingen, Deutschland
+ - name: Max Planck Institute for Biology Tübingen, 72076 Tübingen, Germany
    index: 1
  - name: Research School of Biology, Australian National University, Canberra, Australia
    index: 2
@@ -37,7 +37,7 @@ bibliography: paper.bib
 
 # Summary
 
-Acanthophis is a comprehensive pipeline for the joint analysis of both plant genetic variation and variation in the composition and abundance of plant-associated microbiomes (together, the "hologenome").
+Acanthophis is a comprehensive pipeline for the joint analysis of both host genetic variation and variation in the composition and abundance of host-associated microbiomes (together, the "hologenome").
 Implemented in Snakemake [@koster12_snakemakescalable], Acanthophis handles data from raw FASTQ read files through quality control, alignment of the reads to a plant reference, variant calling, taxonomic classification and quantification of microbes, and metagenome analysis.
 The workflow contains numerous practical optimisations, both to reduce disk space usage and maximise utilisation of computational resources. 
 Acanthophis is available under the Mozilla Public Licence v2 at <https://github.com/kdm9/Acanthophis> as a python package installable from conda or PyPI (`pip install acanthophis`).
@@ -45,14 +45,14 @@ Acanthophis is available under the Mozilla Public Licence v2 at <https://github.
 # Statement of Need
 
 Understanding plant biology benefits from ecosystem-scale analysis of genetic variation, and increasingly demands the characterisation of not only plant genomes but also the genomes of their associated microbes.
-Such analyses are often data intensive, particularly at the scale required for quantitative analyses, i.e. hundreds to thousands of host individuals [@regalado20_combining].
+Such analyses are often data intensive, particularly at the scale required for quantitative analyses, i.e. hundreds to thousands of samples [@regalado20_combining].
 They demand computationally-efficient pipelines that perform both host genotyping and host-associated microbiome characterisation in a consistent, flexible, and reproducible fashion.
 
 Currently, no such unified pipelines exist. Previous pipelines perform only a subset of these tasks (e.g. Snakemake's variant calling pipeline; @koster21_snakemakeworkflows). In addition, most host-aware microbiome analysis pipelines do not allow for genotyping and/or assume an animal host (e.g. Taxprofiler; @yates23_nfcore). Acanthophis has attracted many users, and has been used in peer-reviewed journal articles and preprints (e.g. @murray19_landscapedrivers; @ahrens21_genomicconstraints).
 
 # Components and Features
 
[…]
