### 2024-02-16 #6062 @kdm9
OK @marcosvital, I think the updated manuscript is finally ready, all authors have had a chance to read it and i've updated it with both their and all reviewers comments ([the above discussion](https://github.com/openjournals/joss-reviews/issues/6062#issuecomment-1906472002) notwithstanding).

Pinging also @bricoletc @gbouras13 @abhishektiwari
