### 2024-02-20 kdm9/Acanthophis#9 @bricoletc
@kdm9 this is great! Thanks for answering these; I also re-read the pdf and I'm now checking all the remaining boxes in https://github.com/openjournals/joss-reviews/issues/6062
