### 2024-03-07 #6062 @marcosvital
Dear @bricoletc, @gbouras13 and @abhishektiwari: thank you for all the effort in reviewing this submission.

@kdm9 it seems that we are almost ready to start the acceptance process.  We will need you to deposit the most up to date version on Zenodo (I saw that the last version there is 0.2), so let me know when this is done, ok?
