### 2024-03-07 #6062 @kdm9
@marcosvital Just cut version 1.0.0, and Zenodo is here: https://zenodo.org/records/10795245
