### 2024-03-11 #6062 @marcosvital
@kdm9 can you check the missing DOI for one reference?
