+++ paper/paper.bib
@@ -1,579 +1,520 @@
-
-@article{moran15_hologenomeconcept,
-	title = {The {Hologenome} {Concept}: {Helpful} or {Hollow}?},
-	volume = {13},
-	shorttitle = {The {Hologenome} {Concept}},
-	url = {http://dx.doi.org/10.1371/journal.pbio.1002311},
-	doi = {10.1371/journal.pbio.1002311},
-	abstract = {It is accepted that microbial symbionts are important to the biology of their hosts, but this essay asks what evidence is needed to determine whether symbionts and hosts have coevolved and whether they form higher units of evolutionary selection.},
-	number = {12},
-	urldate = {2015-12-09},
-	journal = {PLoS Biol},
-	author = {Moran, Nancy A. and Sloan, Daniel B.},
-	month = dec,
-	year = {2015},
-	pages = {e1002311},
-	file = {moran15_hologenome__plos_biol__the_hologenome_concept.pdf:/home/kevin/work/bibliography/pdfs/moran15_hologenome__plos_biol__the_hologenome_concept.pdf:application/pdf},
-}
-
-@article{koster12_snakemakescalable,
-	title = {Snakemake --- a scalable bioinformatics workflow engine},
-	volume = {28},
-	issn = {1367-4803, 1460-2059},
-	url = {http://bioinformatics.oxfordjournals.org/content/28/19/2520},
-	doi = {10.1093/bioinformatics/bts480},
-	abstract = {Summary: Snakemake is a workflow engine that provides a readable Python-based workflow definition language and a powerful execution environment that scales from single-core workstations to compute clusters without modifying the workflow. It is the first system to support the use of automatically inferred multiple named wildcards (or variables) in input and output filenames.
-Availability: http://snakemake.googlecode.com.
-Contact: johannes.koester@uni-due.de},
-	language = {en},
-	number = {19},
-	urldate = {2016-03-30},
-	journal = {Bioinformatics},
-	author = {Köster, Johannes and Rahmann, Sven},
-	month = jan,
-	year = {2012},
-	pmid = {22908215},
-	pages = {2520--2522},
-	file = {koster12_snakemake__bioinformatics__snakemake_---_a_scalable_bioinformatics_workflow_engine.pdf:/home/kevin/work/bibliography/pdfs/koster12_snakemake__bioinformatics__snakemake_---_a_scalable_bioinformatics_workflow_engine.pdf:application/pdf},
-}
-
-@article{schubert16_adapterremoval,
-	title = {{AdapterRemoval} v2: rapid adapter trimming, identification, and read merging},
-	volume = {9},
-	issn = {1756-0500},
-	shorttitle = {{AdapterRemoval} v2},
-	url = {http://dx.doi.org/10.1186/s13104-016-1900-2},
-	doi = {10.1186/s13104-016-1900-2},
-	abstract = {As high-throughput sequencing platforms produce longer and longer reads, sequences generated from short inserts, such as those obtained from fossil and degraded material, are increasingly expected to contain adapter sequences. Efficient adapter trimming algorithms are also needed to process the growing amount of data generated per sequencing run.},
-	urldate = {2016-06-07},
-	journal = {BMC Research Notes},
-	author = {Schubert, Mikkel and Lindgreen, Stinus and Orlando, Ludovic},
-	year = {2016},
-	keywords = {High-throughput sequencing, Data pre-processing, Adapter trimming, Adapter identification, Sequence Alignment},
-	pages = {88},
[…]
