### 2024-03-13 867b53d
paper: fix refs
