### 2024-03-21 #6062 @Kevin-Mattheus-Moerman
@kdm9 as AEiC I will now help to process this work for acceptance in JOSS. I have checked this review, your repository, the paper, and the archive. Most seems in order, I only have the below point that needs your attention:

- [ ] Please edit the archive title to match the full paper title.
