### 2024-03-25 #6062 @Kevin-Mattheus-Moerman
@kdm9 Not the zip/tarrball but the actual title and license listed for the archive. See the red highlighted bits in this screenshot:
![temp](https://github.com/openjournals/joss-reviews/assets/8392709/0be14d7d-d1cb-44bc-b3c7-5e46670c0057)
