### 2024-03-25 8d72022
Create CITATION.cff
